Convert zip builtin to class
