+++ Src/IronPython/Runtime/Zip.cs
@@ -1,0 +1,66 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the Apache 2.0 License.
+// See the LICENSE file in the project root for more information.
+
+using System;
+using System.Collections;
+using Microsoft.Scripting.Runtime;
+using IronPython.Runtime.Operations;
+using IronPython.Runtime.Types;
+
+namespace IronPython.Runtime {
+
+    [PythonType("zip")]
+    [Documentation(@"zip(iter1 [,iter2 [...]]) --> zip object
+
+Return a zip object whose .__next__() method returns a tuple where
+the i-th element comes from the i-th iterable argument.  The .__next__()
+method continues until the shortest iterable in the argument sequence
+is exhausted and then it raises StopIteration.")]
+    public class Zip : IEnumerator {
+
+        private readonly IEnumerator[] enumerators;
+        private object current;
+
+        public Zip(CodeContext context, params object[] iters) {
+            if (iters == null) throw PythonOps.TypeError("zip argument #{0} must support iteration", 1);
+
+            enumerators = new IEnumerator[iters.Length];
+            for (var i = 0; i < iters.Length; i++) {
+                if (!PythonOps.TryGetEnumerator(context, iters[i], out enumerators[i]))
+                    throw PythonOps.TypeError("zip argument #{0} must support iteration", i + 1);
+            }
+        }
+
+        public object Current {
+            get {
+                if (current == null) throw new InvalidOperationException();
+                return current;
+            }
+        }
+
+        public bool MoveNext() {
+            if (enumerators.Length > 0 && enumerators[0].MoveNext()) {
+                var res = new object[enumerators.Length];
+                res[0] = enumerators[0].Current;
+
+                for (var i = 1; i < enumerators.Length; i++) {
+                    var enumerator = enumerators[i];
+                    if (!enumerator.MoveNext()) {
+                        current = null;
+                        return false;
+                    }
+
+                    res[i] = enumerator.Current;
+                }
+
+                current = PythonTuple.MakeTuple(res);
+                return true;
+            }
+            current = null;
+            return false;
+        }
+
+        public void Reset() { throw new NotSupportedException(); }
+    }
+}
